requirements.txt
zappa
python-twitter
simplejson
requests
Faker